# docs/swagger.yaml
definitions:
  models.EventData:
    description: If not provided, `videoQuality` & `audioQuality` default to `["720p"]`
      & `["Low"]`, respectively. If provided, first item in the list is event default.
    properties:
      audioQuality:
        example:
        - Low
        - Mid
        - High
        items:
          type: string
        type: array
        uniqueItems: true
      date:
        description: YYYY-MM-DDTHH:MM:SSZ
        example: "2006-01-02T15:04:05Z"
        type: string
      description:
        maxLength: 512
        type: string
      invitees:
        example:
        - example@mail.com
        items:
          type: string
        maxItems: 100
        minItems: 1
        type: array
        uniqueItems: true
      languages:
        example:
        - English
        - French
        items:
          type: string
        minItems: 1
        type: array
        uniqueItems: true
      name:
        description: 'allowed chars: A-Za-z0-9 _-'
        example: A event-Name3_x
        maxLength: 255
        minLength: 1
        type: string
      videoQuality:
        example:
        - 720p
        - 1080p
        - 1440p
        - 2160p
        items:
          type: string
        type: array
        uniqueItems: true
    required:
    - date
    - invitees
    - languages
    - name
    type: object
  models.EventResponseData:
    properties:
      audioQuality:
        example:
        - Low
        - Mid
        - High
        items:
          type: string
        type: array
        uniqueItems: true
      date:
        description: YYYY-MM-DDTHH:MM:SSZ
        example: "2006-01-02T15:04:05Z"
        type: string
      description:
        maxLength: 512
        type: string
      id:
        example: db6bed50-7172-4051-86ab-d1e90705c692
        type: string
      invitees:
        example:
        - example@mail.com
        items:
          type: string
        maxItems: 100
        minItems: 1
        type: array
        uniqueItems: true
      languages:
        example:
        - English
        - French
        items:
          type: string
        minItems: 1
        type: array
        uniqueItems: true
      name:
        description: 'allowed chars: A-Za-z0-9 _-'
        example: A event-Name3_x
        maxLength: 255
        minLength: 1
        type: string
      videoQuality:
        example:
        - 720p
        - 1080p
        - 1440p
        - 2160p
        items:
          type: string
        type: array
        uniqueItems: true
    required:
    - date
    - invitees
    - languages
    - name
    type: object
  models.JsonHealthCheckStatus:
    properties:
      deployDate:
        type: string
      result:
        type: string
      version:
        type: string
    type: object
  weberrors.AppError:
    properties:
      description:
        type: string
      error:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: marek.beck2@gmail.com
    name: Marek Beck
  description: An event management service API in Go using Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: EventHandler API
  version: 1.0.0
paths:
  /event:
    post:
      consumes:
      - application/json
      parameters:
      - description: Event Data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.EventData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EventResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/weberrors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/weberrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/weberrors.AppError'
      summary: Creates event to database
      tags:
      - Event
  /event/{id}:
    delete:
      parameters:
      - description: <token string goes here>
        in: header
        name: API-AUTHENTICATION
        required: true
        type: string
      - description: Event ID (uuid)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/weberrors.AppError'
      summary: Delete event from database
      tags:
      - Event
    get:
      parameters:
      - description: Event ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EventResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/weberrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/weberrors.AppError'
      summary: Retrieves event from database
      tags:
      - Event
  /healthcheck:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JsonHealthCheckStatus'
      summary: Checks health of this service
      tags:
      - Health check
swagger: "2.0"